% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{birthday function}
\usage{
birthday(k)
}
\arguments{
\item{k}{the number of people in the room}
}
\value{
Probability
}
\description{
birthday function
}
\examples{
birthday(20)
}
