% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations to run}

\item{x}{function you want to run}

\item{fun}{what type of statistic you want to see}

\item{alpha}{1 - the confidence interval}

\item{cx}{value}

\item{...}{accepts any further data}
}
\value{
a histogram of the chosen statistic with confidence intervals
}
\description{
myboot2
}
