% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n = 10, iter = 10000)
}
\arguments{
\item{n}{size wanted}

\item{iter}{number of iterations to run}
}
\value{
a histogram of w
}
\description{
myclt
}
\examples{
\dontrun{w=myclt(n=10,iter=10000)}
}
