% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets_fun.R
\name{ntickets}
\alias{ntickets}
\title{ntickets_fun}
\usage{
ntickets(N = 400, g = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of Seats}

\item{g}{gamma}

\item{p}{percent to show}
}
\value{
two graphs, discrete and continuous of n
}
\description{
ntickets_fun
}
