% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{myf}
\usage{
myf(y, r, p)
}
\arguments{
\item{y}{Total number of trials (integer)}

\item{r}{Number of successes (integer)}

\item{p}{Probability of success in each trial (0 < p < 1)}
}
\description{
myf
}
