% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynfunction.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}

\item{a}{stopping point}
}
\description{
myncurve
}
